Fritz tests cleanup
