+++ tests/components/fritz/test_config_flow.py
@@ -3,9 +3,7 @@ import dataclasses
 from unittest.mock import patch
 
 from fritzconnection.core.exceptions import FritzConnectionException, FritzSecurityError
-import pytest
 
-from homeassistant.components import ssdp
 from homeassistant.components.device_tracker.const import (
     CONF_CONSIDER_HOME,
     DEFAULT_CONSIDER_HOME,
@@ -16,9 +14,9 @@ from homeassistant.components.fritz.const import (
     ERROR_CANNOT_CONNECT,
     ERROR_UNKNOWN,
 )
-from homeassistant.components.ssdp import ATTR_UPNP_FRIENDLY_NAME, ATTR_UPNP_UDN
+from homeassistant.components.ssdp import ATTR_UPNP_UDN
 from homeassistant.config_entries import SOURCE_REAUTH, SOURCE_SSDP, SOURCE_USER
-from homeassistant.const import CONF_DEVICES, CONF_HOST, CONF_PASSWORD, CONF_USERNAME
+from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import (
     RESULT_TYPE_ABORT,
@@ -26,42 +24,15 @@ from homeassistant.data_entry_flow import (
     RESULT_TYPE_FORM,
 )
 
-from . import MOCK_CONFIG, FritzConnectionMock
-
-from tests.common import MockConfigEntry
-
-ATTR_HOST = "host"
-ATTR_NEW_SERIAL_NUMBER = "NewSerialNumber"
-
-MOCK_HOST = "fake_host"
-MOCK_IP = "192.168.178.1"
-MOCK_SERIAL_NUMBER = "fake_serial_number"
-MOCK_FIRMWARE_INFO = [True, "1.1.1"]
-
-MOCK_USER_DATA = MOCK_CONFIG[DOMAIN][CONF_DEVICES][0]
-MOCK_DEVICE_INFO = {
-    ATTR_HOST: MOCK_HOST,
-    ATTR_NEW_SERIAL_NUMBER: MOCK_SERIAL_NUMBER,
-}
-MOCK_SSDP_DATA = ssdp.SsdpServiceInfo(
-    ssdp_usn="mock_usn",
-    ssdp_st="mock_st",
-    ssdp_location=f"https://{MOCK_IP}:12345/test",
-    upnp={
-        ATTR_UPNP_FRIENDLY_NAME: "fake_name",
-        ATTR_UPNP_UDN: "uuid:only-a-test",
-    },
+from .const import (
+    MOCK_FIRMWARE_INFO,
+    MOCK_IP,
+    MOCK_REQUEST,
+    MOCK_SSDP_DATA,
+    MOCK_USER_DATA,
 )
 
-MOCK_REQUEST = b'<?xml version="1.0" encoding="utf-8"?><SessionInfo><SID>xxxxxxxxxxxxxxxx</SID><Challenge>xxxxxxxx</Challenge><BlockTime>0</BlockTime><Rights><Name>Dial</Name><Access>2</Access><Name>App</Name><Access>2</Access><Name>HomeAuto</Name><Access>2</Access><Name>BoxAdmin</Name><Access>2</Access><Name>Phone</Name><Access>2</Access><Name>NAS</Name><Access>2</Access></Rights><Users><User last="1">FakeFritzUser</User></Users></SessionInfo>\n'
-
-
-@pytest.fixture()
-def fc_class_mock():
-    """Fixture that sets up a mocked FritzConnection class."""
-    with patch("fritzconnection.FritzConnection", autospec=True) as result:
-        result.return_value = FritzConnectionMock()
-        yield result
+from tests.common import MockConfigEntry
 
 
 async def test_user(hass: HomeAssistant, fc_class_mock, mock_get_source_ip):
